HSEARCH-1316 Move Spring JTA ITs to a subpackage of org.hibernate.search.test.integration.spring
