+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/spring/jta/JtaAndSpringIT.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring;
+package org.hibernate.search.test.integration.spring.jta;
 
 import javax.inject.Inject;
 
-import org.hibernate.search.test.integration.jtaspring.dao.SnertDAO;
-import org.hibernate.search.test.integration.jtaspring.entity.Snert;
+import org.hibernate.search.test.integration.spring.jta.dao.SnertDAO;
+import org.hibernate.search.test.integration.spring.jta.entity.Snert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
