+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/spring/jta/dao/MuffinDAO.java
@@ -4,37 +4,32 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring.dao;
+package org.hibernate.search.test.integration.spring.jta.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
-import org.hibernate.search.test.integration.jtaspring.entity.Box;
+import org.hibernate.search.test.integration.spring.jta.entity.Muffin;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public class BoxDAO {
+public class MuffinDAO {
 	@PersistenceContext
 	private EntityManager entityManager;
 
 	@Transactional
-	public void persist(Box box) {
-		entityManager.persist( box );
+	public void persist(Muffin muffin) {
+		entityManager.persist( muffin );
 	}
 
 	@Transactional
-	public Box merge(Box box) {
-		Box result = entityManager.merge( box );
-		return result;
+	public void merge(Muffin muffin) {
+		entityManager.merge( muffin );
 	}
 
 	@Transactional
-	public void remove(Box box) {
-		entityManager.remove( box );
-	}
-
-	public EntityManager getEntityManager() {
-		return entityManager;
+	public void remove(Muffin muffin) {
+		entityManager.remove( muffin );
 	}
 }
