+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/spring/jta/dao/SnertDAO.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring.dao;
+package org.hibernate.search.test.integration.spring.jta.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
-import org.hibernate.search.test.integration.jtaspring.entity.Snert;
+import org.hibernate.search.test.integration.spring.jta.entity.Snert;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
