+++ integrationtest/spring/src/test/java/org/hibernate/search/test/integration/spring/jta/entity/Snert.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.integration.jtaspring.entity;
+package org.hibernate.search.test.integration.spring.jta.entity;
 
 import java.util.Date;
 import javax.persistence.Column;
